refactor(file-router): remove & extract non-public API elements
