+++ packages/ts/hilla-file-router/src/types.d.ts
@@ -9,14 +9,9 @@ export type ViewConfig = Readonly<{
   title?: string;
 
   /**
-   * A map of route parameters and their types.
-   */
-  params?: Readonly<Record<string, RouteParamType>>;
-
-  /**
    * Same as in the explicit React Router configuration.
    */
-  rolesAllowed?: string[];
+  rolesAllowed?: readonly string[];
 
   /**
    * Allows overriding the route path configuration. Uses the same syntax as
@@ -24,21 +19,6 @@ export type ViewConfig = Readonly<{
    * that conflicts with the file name conventions, e.g. /foo/index.
    */
   route?: string;
-
-  /**
-   * Controls whether the view implementation will be lazy loaded the first time
-   * it's used or always included in the bundle. If set to undefined (which is
-   * the default), views mapped to / and /login will be eager and any other view
-   * will be lazy (this is in sync with defaults in Flow)
-   */
-  lazy?: boolean;
-
-  /**
-   * If set to false, then the route will not be registered with React Router,
-   * but it will still be included in the main menu and used to configure
-   * Spring Security
-   */
-  register?: boolean;
 
   menu?: Readonly<{
     /**
@@ -80,4 +60,41 @@ export type AgnosticRoute<C = unknown> = Readonly<{
   path: string;
   module?: RouteModule<C>;
   children?: ReadonlyArray<AgnosticRoute<T>>;
+}>;
+
+/**
+ * Internal type used for server communication and menu building. It extends the
+ * view configuration with the route parameters.
+ *
+ * @internal
+ */
+export type ServerViewConfig = Readonly<{
+  /**
+   * A map of route parameters and their types.
+   */
+  params?: Readonly<Record<string, RouteParamType>>;
+}> &
+  ViewConfig;
+
+export type VaadinServer = Readonly<{
+  views: Readonly<Record<string, ServerViewConfig>>;
+}>;
+
+export type VaadinObject = Readonly<{
+  server?: VaadinServer;
+}>;
+
+declare global {
+  interface Window {
+    Vaadin?: VaadinObject;
+  }
+}
+
+/**
+ * A menu item used in for building the navigation menu.
+ */
+export type MenuItem = Readonly<{
+  to: string;
+  icon?: string;
+  title?: string;
 }>;
