+++ packages/ts/hilla-file-router/src/vite-plugin/createViewConfigJson.ts
@@ -3,7 +3,7 @@ import { Script } from 'node:vm';
 import ts, { type Node } from 'typescript';
 import { convertComponentNameToTitle } from '../shared/convertComponentNameToTitle.js';
 import traverse from '../shared/traverse.js';
-import type { ViewConfig } from '../types.js';
+import type { ServerViewConfig, ViewConfig } from '../types.js';
 import type { RouteMeta } from './collectRoutesFromFS.js';
 import { convertFSRouteSegmentToURLPatternFormat, extractParameterFromRouteSegment } from './utils.js';
 
@@ -30,7 +30,7 @@ export default async function createViewConfigJson(views: RouteMeta): Promise<st
   const res = await Promise.all(
     Array.from(traverse(views), async (branch) => {
       const configs = await Promise.all(
-        branch.map(async ({ path, file, layout }): Promise<[string, ViewConfig]> => {
+        branch.map(async ({ path, file, layout }): Promise<[string, ServerViewConfig]> => {
           if (!file && !layout) {
             return [
               convertFSRouteSegmentToURLPatternFormat(path),
